adding force clear costmap state; all nav recovery states put in the same file
